Reorg connection (#891)
* change resolver code based on new gql type

* comments and cleanup

* break up main entry point into two modules, one for querying and one for field processing

* rename network config type, remove supported aggs from type

* type cleanup

* only send network queries with original query fields

* move gql health check into module

* cleanup, renaming types, removing redundant code

* add type

* fix merged conflict

* adds full stop. fixes comment typo

---------
