+++ modules/server/src/network/types/aggregations.ts
@@ -1,0 +1,15 @@
+/*
+ * Typescript types for aggregations
+ */
+
+export type Bucket = {
+	doc_count: number;
+	key: string;
+};
+
+export type Aggregations = {
+	bucket_count: number;
+	buckets: Bucket[];
+};
+
+export type NumericAggregations = {};
