+++ modules/server/src/network/types/setup.ts
@@ -1,0 +1,10 @@
+/*
+ * Setup of network types
+ */
+
+export type NetworkConfig = {
+	graphqlUrl: string;
+	documentType: string;
+	documentName: string;
+	displayName: string;
+};
